+++ Source/emergeTasks/Applet.cpp
@@ -207,7 +207,7 @@ void Applet::UpdateIcons()
 // Returns:	LRESULT
 // Purpose:	Paints the icons on the calling window
 //----  --------------------------------------------------------------------------------------------------------
-bool Applet::PaintItem(HDC hdc, UINT index, int x, int y, RECT rect)
+bool Applet::PaintItem(HDC hdc, size_t index, int x, int y, RECT rect)
 {
 
   TaskVector::iterator iter = taskList.begin() + index;
